# contracts/openapi/v1.yaml
openapi: "3.1.0"
info:
  title: Agent Framework Orchestrator API
  version: "1.0.0"
  description: |
    REST API for the Agent Framework control plane orchestrator.
    Accepts natural language specifications, decomposes them into
    executable tasks, and coordinates specialized AI agents.
  contact:
    name: Agent Framework Team
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://agent-framework-dev.azurecontainerapps.io
    description: Development environment
paths:
  /api/v1/plans:
    post:
      operationId: createPlan
      summary: Create and start a new execution plan
      description: |
        Accepts a natural language specification, decomposes it into
        an ordered list of tasks using the planner agent, and starts
        dispatching tasks to specialized workers.
      tags:
        - Plans
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "v1.components.yaml#/components/schemas/PlanRequest"
            examples:
              simpleApi:
                summary: Simple REST API
                value:
                  spec: "Build a REST API for user management with JWT authentication"
              fullStack:
                summary: Full-stack application
                value:
                  spec: "Build a todo application with React frontend and Spring Boot backend"
                  constraints:
                    database: postgresql
                    auth: jwt
                  tags:
                    - fullstack
                    - crud
      responses:
        "202":
          description: Plan created and execution started
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/PlanResponse"
        "400":
          description: Invalid request (spec too short, malformed JSON)
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
        "500":
          description: Internal error (planner failure, Service Bus unavailable)
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
  /api/v1/plans/{planId}:
    get:
      operationId: getPlan
      summary: Get plan details with all items
      tags:
        - Plans
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
      responses:
        "200":
          description: Plan details
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/PlanResponse"
        "404":
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
  /api/v1/plans/{planId}/quality-gate:
    get:
      operationId: getQualityGateReport
      summary: Get quality gate report for a completed plan
      tags:
        - Quality Gate
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
      responses:
        "200":
          description: Quality gate report
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/QualityGateReportResponse"
        "404":
          description: Plan or report not found
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
  /api/v1/plans/{planId}/items/{itemId}/retry:
    post:
      operationId: retryItem
      summary: Retry a failed plan item
      description: |
        Transitions the item from FAILED back to WAITING and triggers a new dispatch wave.
        If the parent plan was FAILED, it is also reopened as RUNNING.
      tags:
        - Plans
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
        - $ref: "v1.components.yaml#/components/parameters/ItemId"
      responses:
        "202":
          description: Retry accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  itemId:
                    type: string
                    format: uuid
                  planId:
                    type: string
                    format: uuid
        "400":
          description: Item is not in FAILED state
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
  /api/v1/plans/{planId}/items/{itemId}/attempts:
    get:
      operationId: getItemAttempts
      summary: Get dispatch history for a plan item
      description: Returns all dispatch attempts for the given item, ordered by attempt number.
      tags:
        - Plans
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
        - $ref: "v1.components.yaml#/components/parameters/ItemId"
      responses:
        "200":
          description: List of dispatch attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "v1.components.yaml#/components/schemas/DispatchAttemptResponse"
  /api/v1/plans/{planId}/snapshots:
    get:
      operationId: listSnapshots
      summary: List all snapshots for a plan
      tags:
        - Plans
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
      responses:
        "200":
          description: List of plan snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "v1.components.yaml#/components/schemas/PlanSnapshotResponse"
  /api/v1/plans/{planId}/restore/{snapshotId}:
    post:
      operationId: restoreSnapshot
      summary: Restore a plan to a previous snapshot
      tags:
        - Plans
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
        - $ref: "v1.components.yaml#/components/parameters/SnapshotId"
      responses:
        "200":
          description: Plan restored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  planId:
                    type: string
                    format: uuid
                  snapshotId:
                    type: string
                    format: uuid
        "400":
          description: Restore failed
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
  /api/v1/plans/{planId}/resume:
    post:
      operationId: resumePlan
      summary: Resume a paused plan
      description: |
        Transitions the plan from PAUSED back to RUNNING and re-triggers
        the dispatch wave. Auto-retry failures that exceeded the threshold
        pause the plan automatically; this endpoint allows manual resumption.
      tags:
        - Plans
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
      responses:
        "202":
          description: Plan resumed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: resuming
                  planId:
                    type: string
                    format: uuid
        "400":
          description: Plan is not in PAUSED state
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
  /api/v1/plans/{planId}/items/{itemId}/issue-snapshot:
    put:
      operationId: saveIssueSnapshot
      summary: Save issue tracker snapshot on a plan item
      description: |
        Internal endpoint called by the TASK_MANAGER worker to persist the
        JSON snapshot fetched from the external issue tracker. Not intended
        for external callers.
      tags:
        - Plans
        - Internal
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
        - $ref: "v1.components.yaml#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - snapshotJson
              properties:
                snapshotJson:
                  type: string
                  description: JSON string of the issue snapshot from the tracker
      responses:
        "200":
          description: Snapshot saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: saved
                  itemId:
                    type: string
                    format: uuid
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
  /api/v1/plans/{planId}/items/{itemId}/approve:
    post:
      operationId: approveItem
      summary: Approve a task held in AWAITING_APPROVAL
      description: |
        Approves a task that was classified as high-risk (CRITICAL) by the
        HOOK_MANAGER. Transitions AWAITING_APPROVAL → WAITING and triggers
        a new dispatch wave.
      tags:
        - Plans
        - Approval
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
        - $ref: "v1.components.yaml#/components/parameters/ItemId"
      responses:
        "200":
          description: Item approved and re-queued for dispatch
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: approved
                  itemId:
                    type: string
                    format: uuid
        "400":
          description: Item is not in AWAITING_APPROVAL state
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
  /api/v1/plans/{planId}/items/{itemId}/reject:
    post:
      operationId: rejectItem
      summary: Reject a task held in AWAITING_APPROVAL
      description: |
        Rejects a task that was classified as high-risk. Transitions
        AWAITING_APPROVAL → FAILED with the provided reason.
      tags:
        - Plans
        - Approval
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
        - $ref: "v1.components.yaml#/components/parameters/ItemId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for rejection (defaults to "Rejected by human reviewer")
      responses:
        "200":
          description: Item rejected
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: rejected
                  itemId:
                    type: string
                    format: uuid
        "400":
          description: Item is not in AWAITING_APPROVAL state
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
  /api/v1/plans/{planId}/items/{itemId}/compensate:
    post:
      operationId: compensateItem
      summary: Create a compensation task for a plan item
      description: |
        Creates a COMPENSATOR_MANAGER task to undo the effects of the
        specified item (e.g. git revert). The compensation task is added
        to the plan and dispatched immediately.
      tags:
        - Plans
        - Compensation
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
        - $ref: "v1.components.yaml#/components/parameters/ItemId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for compensation (defaults to "Manual compensation requested")
      responses:
        "202":
          description: Compensation task created and dispatched
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: compensation_started
                  compensationItemId:
                    type: string
                    format: uuid
                  compensationKey:
                    type: string
                    description: Task key of the compensation task (e.g. "CP-001")
                  planId:
                    type: string
                    format: uuid
        "400":
          description: Compensation not possible for this item
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
  /api/v1/plans/{planId}/events:
    get:
      operationId: streamPlanEvents
      summary: Subscribe to real-time plan events via SSE
      description: |
        Returns a Server-Sent Events stream of plan state changes.
        On connection, all past events for the plan are replayed (late-join replay),
        then live events are streamed as they occur.
        The connection stays open for up to 5 minutes (clients should reconnect after timeout).
      tags:
        - Plans
        - Events
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
      responses:
        "200":
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Each SSE event has `data` containing a JSON PlanEvent:
                  {"planId":"...","itemId":"...","eventType":"TASK_COMPLETED","payload":"...","occurredAt":"..."}
  /api/v1/plans/{planId}/graph:
    get:
      operationId: getPlanGraph
      summary: Get task DAG as Mermaid diagram or JSON
      description: |
        Returns the task dependency graph (DAG) of a plan. Each node includes
        taskKey, workerType, status, duration, and token usage (when available).
        Edges represent dependsOn relationships with inferred labels.

        The Mermaid output can be embedded directly in Gitea/GitHub Markdown
        inside a ```mermaid``` code block.
      tags:
        - Plans
        - Graph
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum:
              - mermaid
              - json
            default: mermaid
          description: Output format (mermaid for text diagram, json for adjacency list)
      responses:
        "200":
          description: Plan graph
          content:
            text/plain:
              schema:
                type: string
                description: Mermaid graph diagram (when format=mermaid)
            application/json:
              schema:
                type: string
                description: JSON adjacency list (when format=json)
        "404":
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"
  /api/v1/plans/{planId}/council-report:
    get:
      operationId: getCouncilReport
      summary: Get the pre-planning council report for a plan
      description: |
        Returns the JSON CouncilReport produced by the pre-planning council session,
        or 204 No Content if council was disabled or the plan predates the council feature.
      tags:
        - Plans
        - Council
      parameters:
        - $ref: "v1.components.yaml#/components/parameters/PlanId"
      responses:
        "200":
          description: Council report JSON
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/CouncilReport"
        "204":
          description: No council report (council disabled or plan predates the feature)
        "404":
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: "v1.components.yaml#/components/schemas/ErrorResponse"

  # ── Reward System ─────────────────────────────────────────────────────
...